Updated to latest version
--- Install ---
* Added new default permissions, and added a
registration page by default to page settings.

-- Resources --
* Now
properly support full searches - pagination, number of matching items
regardless of limit, etc.
* Can now get the collections that a resource
belongs to.

-- Upload Handler --
* Now create folders that dont exist
on upload.

-- Media & FroteEndResourceHandler --
* Added new
functionalities - createFolder, getImageGalleries.
* Now supports full
searches, like Resources.
* Fixes and stuff.
* Many functions now create
folders that dont exist.

-- Helper Functions --
* A few fixes

-- Front
End --
* Slightly expanded, and made some small fixes in ezAlert.
*
Small fixes in utils.JS, alert function slightly expanded, getLuminance
added.

+++ api/mediaAPI_fragments/uploadImages_execution.php
@@ -19,7 +19,7 @@ foreach($inputs['items'] as $uploadName => $item){
 $result = $UploadHandler->handleUploadedImage(
     $uploadNames,
     [
-        'test'=>false,
+        'test'=>$test,
         'verbose'=>$test,
         'overwrite'=>$inputs['overwrite'] && !$test,
         'imageQualityPercentage'=>$inputs['imageQualityPercentage'],
@@ -67,7 +67,7 @@ foreach($result as $uploadName => $res){
                 $meta['name'] = $name;
             if($caption)
                 $meta['caption'] = $caption;
-            $meta = [json_encode($meta),'STRING'];
+            $meta = json_encode($meta);
         }
             else
                 $meta = null;
@@ -121,7 +121,6 @@ foreach($updateDB as $DBName => $resultCode){
 
 //Update the gallery if requested
 if($imagesToAddToGallery != []){
-
     $updateDB =  $FrontEndResourceHandler->addImagesToGallery(
         $imagesToAddToGallery,
         $inputs['gallery'],
@@ -129,7 +128,7 @@ if($imagesToAddToGallery != []){
     );
 
     foreach($updateDB as $DBName => $resultCode){
-        if($resultCode === 0 ){
+        if($resultCode === 0 || $test){
             //If this is a test, always delete
             if(!$test)
                 unset($deleteLocalFiles[$dbToUploadMap[$DBName]]);
@@ -141,6 +140,10 @@ if($imagesToAddToGallery != []){
 }
 
 //Delete files that we still need to delete
-foreach($deleteLocalFiles as $uploadName => $addr)
-    unlink($addr);
+foreach($deleteLocalFiles as $uploadName => $addr){
+    if(!$test)
+        unlink($addr);
+    else
+        echo 'Deleting file at '.$addr.EOL;
+}
